Implement complete remote data source integration with A2SV API and full test coverage

+++ mobile/AbdrehimMisbah/ecommerce_ui_replication/lib/features/product/data/models/product_model.dart
@@ -22,7 +22,7 @@ class ProductModel extends Product {
   // 🧠 Important: fromJson converts Map → ProductModel
   factory ProductModel.fromJson(Map<String, dynamic> json) {
     return ProductModel(
-      id: json['id'] as int,
+      id: json['id'] ,
       name: json['name'],
       imageUrl: json['imageUrl'],
       price: (json['price'] as num).toDouble(), // handles 1 or 1.0
